requirements.txt
altgraph==0.17.2
ansicolors==1.1.8
cefpython3==66.1
Cython==0.29.24
Events==0.4
future==0.18.2
Jinja2==3.0.2
MarkupSafe==2.0.1
pefile==2021.9.3
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32-ctypes==0.2.0
